backend/requirements.txt
Django>=3.0,<4.0
django-cleanup>=5.1.0
djangorestframework>=3.12.2
Pillow>=8.0.1
gunicorn>=20.0.4
psycopg2-binary>=2.8
django-axes>=5.9.1
djangorestframework-simplejwt>=4.1.2
djoser